Improve `User.email` handling for new accounts
`auth.info.email` will return `nil` when the Rocket.Chat user's email wasn't confirmed. Depending on the RC configuration it is not possible to confirm the email. We use the unconfirmed email as a fallback now. This should be fixed on the rocketchat provider side. We also may or may not need an email confirmed field in our user model.