Refactor task-from-preset creation mode
GlobalID works well when we want to derive the preset values from a resource that is persisted to the database, e.g. another task. But once we want to have presets from other resources, e.g. a project, to set preset values on the task, then GlobalID isn't suited anymore.

This refactoring implements url-params based presets and sets up a new route.