Add `modal` section to application layout
For some components we want to render a modal within the component, but it should appear elsewhere in the DOM, e.g. when the modal contains a `<form>` and the component is part of a `<form>` itself. Now, the modal can be surrounded with a `content_for :modal` block and will appear close to the end of the `<body>` element.