Fix mobile navigation by partial revert
Removing `flex-wrap` prevents the navigational elements from wrapping to a second line on very small viewports. This was desired behaviour. However, this would break the mobile menu, because it relies on being able to wrap.

The mobile menu has priority, so for now `flex-wrap` was added back in until a fix has been figured out.