Improve deeplink handling
Previously we let https://go.rocket.chat handle deeplinks with web link fallbacks. This wasn't an issue because we could use anonymous user ids. Sadly, these anonymous links only work when users already established a PM in the past. Otherwise, an error page appears. To mitigate we have to build links with the username. The username potentially contains personal data. So we cannot use https://go.rocket.chat anymore to resolve the links.

In general, deeplinks are an extensive topic. We experimented with implementations that try to open `rocketchat://` and gracefully downgrade to `https://` when the client doesn't have the app installed. This usually works by detecting events when the web application loses focus, such as "blur". Triggering a "blur" event would mean the app handler successfully opened the app. However, these implementations are fragile and unreliable. For example, some browsers open a dialog asking how to handle the unknown protocol, also leading to lost focus.

Thus, a simpler approach was adopted:

* Detect if the client is on Android or iOS
* Yes, then open the `rocketchat://` handler
* No, then just open a regular `https://` link

This should work well, because most users should have the mobile client installed. On desktop most users probably use the web browser.

If this turns out to be a problem, the implementation could shift to a Rails helper PORO. Detection could be done with the `device_detector` gem. We would users allow to adjust the behaviour through a setting in their profile.