Introduce I18n
Bootstraps the whole project with I18n in German (complete) and English (possibly incomplete).

Changes naming of the project entity
- Projekt -> Thema in German
- Project -> Initiative in English
In the code base we will continue to use project consistently for this entity. Might change it later when the application is more mature.

Some refactoring along the way. Leads to an unreasonable large commit (sorry), but it was a good opportunity because almost every file had to be touched anyway.