Refactor `task_applications` primary key
The `noticed` gem for sending notifications requires a single integer PK. The `task_applications` table was originally designed with a composite PK of `task_id` and `user_id`.

This code migrates the structure and data to a regular auto-increment `id` PK and adds a unique constraint on the composite of `task_id` and `user_id`.